requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
filelock==3.25.2
fsspec==2026.2.0
gpt4all==2.8.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.3
openai==2.29.0
packaging @ file:///home/task_176104878999627/conda-bld/packaging_1761049078606/work
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
setuptools==80.10.2
shellingham==1.5.4
sniffio==1.3.1
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
